chore: separation of business logic at hook
