+++ src/app/login/styles.ts
@@ -12,7 +12,8 @@ export const Main = styled('main')`
 `;
 
 export const Wrapper = styled('div')`
-  width: 25rem;
+  width: 90%;
+  max-width: 25rem;
   display: flex;
   align-items: center;
   gap: 1rem;
@@ -20,6 +21,7 @@ export const Wrapper = styled('div')`
   box-shadow: 0 0 0.625rem 0.0625rem rgba(0, 0, 0, 0.34);
 
   ${({ theme }) => theme.breakpoints.up('sm')} {
+    max-width: none;
     width: 37.5rem;
     height: 25rem;
   }
@@ -29,18 +31,27 @@ export const SideLeft = styled('div')`
   width: 40%;
   height: 100%;
   background-color: ${({ theme }) => theme.palette.grey[300]};
+  display: none;
 
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  ${({ theme }) => theme.breakpoints.up('sm')} {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
 `;
 
 export const Form = styled('form')`
   padding: ${spacing.S10};
+  padding-right: ${spacing.S20};
+  margin: 0 ${spacing.S20};
   display: flex;
   flex-direction: column;
   gap: 1rem;
-  width: 60%;
+  width: 100%;
+
+  ${({ theme }) => theme.breakpoints.up('sm')} {
+    width: 60%;
+  }
 `;
 
 export const FooterForm = styled('div')`
